Tuning Fontconfig - correct the details for KDE.
I think I had intended to mention /etc/fonts/local.conf as the
all-user override of KDE settings, but my text was garbled and
it appears that a system-wide override of fonts used by KDE
applications is no-longer possible.

Also add a link to 'Items which can override Fontconfig' as the
second paragraph of the page, because on any modern desktop many
of the settings for Sans, Serif and Monospace might be ignored.