webkitgtk: Disable WebKitWebDriver for GTK-3 build
It'll be overwritten by the GTK-4 build so we can save some build time.
And if we consider the GTK-3 build and the GTK-4 build as two packages
it's bad to allow a part of one package overwritten by another.

To be honest we even don't need WebKitWebDriver for the GTK-4 build as
no BLFS packages ever need it, but AFAIK we don't aggressively prune
packages so just keep it for GTK-4.