colord: Patch to fix two issues
The first issue only crashes the daemon when a sqlite error happens, so
it's mostly latent.  The second issue causes /var/lib/colord/*.db opened
as read-only, it won't cause an issue until colord needs to modify the
content of these files.  Note that for a fresh installation (w/o
/var/lib/colord/*.db) colord will immediately fail to start creating
these files.