rust: Remove the hack for non-SSE2 system
We are going to unsupport BLFS on 32-bit x86, and the sed command just
doesn't work for recent Rust releases.

The upstream says you should configure for i586-unknown-linux-gnu (5
instead of 6) if the system does not support SSE2 but I've not tested it
at all.  It's unlikely anyone will ever test it (who'll build LFS, LLVM,
and then Rustc on a Pentium III?)