openssh: Explicitly unsupport changing ListenAddress when using BLFS configuration
I've not seen any distro supporting this.

Changing ListenAddress is already a very bad idea: for e.g. if a
repairman replaced the ethernet cable, a sshd with custom ListenAddress
will just exit and you cannot access the machine remotely anymore...
Adding Wanted=network-online.target into sshd.service will also make
things worse: when the system boots if the network connection is in a
degraded status (for e.g. one NIC is ok but another is not) systemd
won't even try to start sshd and again you cannot access the machine
remotely.

So let's explicitly unsupport doing so.  If you want to limit the access
of ssh port, use iptables.