Patch gtk+2 to build with gcc14.
The patch includes this week's gtk3 fix, but I am not convinced
that part is actually a vulnerability on recent BLFS. Using the
PoC and the ENVVAR and starting gimp-2.10 from a prepared directory
the PoC failed on gtk2 built with only the gcc14 fixes.

As noted in the comment I added to the patch, the reporter thinks
exploitation on GTK2 is unlikely.

Testing epdfview, the PoC succeeded.