firefox, thunderbird: Enable rust SIMD again
The shipped encoding_rs crate has been ported from the broken
packed_simd to portable_simd.  The latter is still a nightly feature but
the building system has some internal magic to enable nightly features
for our stable rustc build.

The danger of using nightly feature is it'll be more likely to be broken
with the future rustc releases.  But we can then revert this change if
it happens anyway.