Update to BIND-9.20.0.
There's a lot of changes that went in here. The tests require the
hypothesis python module, so I've added in a set of commands to setup a
virtual environment and then run the tests from there.

Two libraries were removed, which required changes over in BIND
Utilities to instructions and to the installed files in BIND

The DLZ support has been removed sine BIND-9.18.0 (20 releases ago), so
I've removed the option for that from Command Explanations and dropped
the dependencies.