uefi/grub-setup: Don't insmod all_video
In a QEMU VM, "insmod all_video" in grub will cause Linux kernel
simpledrm driver to misbehave and display the boot messages (or anything
printed before bochs.ko is loaded) like a 3 week old slice of pizza in
the fridge.

I've not observed this "effect" on real hardware yet.  I guess the
reason is GRUB has a "native" driver for the QEMU virtual GPU (bochs)
and that driver changes video setting w/o letting the UEFI know, then
UEFI passes outdated info to the kernel and sh*t happens.  OTOH on real
hardware GRUB does not have any driver for AMD/NVIDIA/Intel, so it has
to use the UEFI-based driver and there's nothing messing with the GPU
in a way unknown to the UEFI.

So to make it work in QEMU we can just load UEFI-based video drivers,
instead of all video drivers.  And this works for me.