bluefish: Disable the infbrowser plugin instead of building a broken one
The error message from GCC 14 indicates a real problem: the NanoHTTP
functions are removed from libxml >= 2.13.0.  Manually injecting a
prototype just papers over the issue as the function just does not exist
in libxml2.so:

    $ objdump -T /usr/lib/libxml2.so | grep xmlNanoHTTPFetch \
      || echo nothing
    nothing

Thus we are just building a plugin which isn't loadable at runtime at
all.  Note that when linking a shared object the linker does not mandate
the referenced symbols to exist but dlopen() a shared object referencing
a nonexist symbol will cause an "undefined symbol" error.

So it's better not to build this plugin at all, instead of building and
installing a completely broken one.