libgdata: The tests do not work with the latest uhttpmock.
The problem here is quite involved - libgdata still uses soup2, while
uhttpmock uses soup3. Meson originally won't find uhttpmock-0.0 (new
version is 1.0), but if you apply a sed to allow it to build, all of the
tests will fail because it mixes soup2 and soup3 symbols in the same
program

For now let's add a comment to the page, note that the package doesn't
come with a functional test suite, and then comment out the uhttpmock
dependency. Once libgdata is updated to use libsoup3, let's review this
again. Note that since it's using soup2 we don't currnetly have Google
Drive integration in gvfs, *even if the option is passed to gvfs*