clisp: Don't decrease stack hard limit
In bash "ulimit -s 16384" sets both the soft and hard limits to 16 MiB
(as long as the hard limit is not smaller than 16 MiB).  But the default
hard limit is infinite, thus we are decreasing the hard limit.  Add "-S"
so it only changes the soft limit (the default is 8 MiB and the
configure script says it's not enough).