Archive qt5.
There are a large number of packages that needed to be changed to
remove qt5, but most are harmless.  One package that is affected
is vlc where the gui interface is no longer available.  Videos
can still be played from the command line and the kf6 dependency
on vlc is still satisfied.

If necessary qt5 can be restored for vlc.

Also three other files are archived that are no longer needed.