libblockdev: Add new dependencies for the SMART plugin
Looking at
https://github.com/storaged-project/libblockdev/blob/3.2.0/NEWS.rst,
we'll notice that a new plugin was added for SMART support on hard
drives and SSDs. The default plugin is based on libatasmart, so let's
add that to the Recommended dependencies list. The experimental plugin
is based on smartmontools and uses the 'smartctl' executable at runtime,
so we'll add that (and the json-glib dependency that it requires).