askhelp: Discourage and encourage scripting
The subject seems irony.  We should clearly say scripting is deviating
from the book because there are too many cases where people fail to
notice a subtle difference between the script and the book.  But if you
must deviate anyway (no matter scripting or not), it's better to have a
script to tell precisely what the deviation is.

I still believe scripting does not make too much sense for building
BLFS, but having the script is at least better than guessing what the
script contains.