firmware: Remove the fake info about reclocking
The firmware released by NVIDIA supports reclocking.  It's just
experimental for old cards (Kepler and earler, only available via
debugfs) and not implemented for new cards by Nouveau yet.

This is indeed an issue for Tesla through Pascal because the default
clock speed is too low and causing performance issues.  But anyway it's
not a firmware bug, and it shouldn't be documented here.