kernel-config: mesa: Just enable GSP for Turing/Ampere
Without GSP it's reported the performance of Turing/Ampere is
"horrible."  Arch has also enabled GSP for them by default.

And generally it's better to use a code path similar to a newer GPU
instead of one similar to an older GPU if both are available because the
former is often better maintained (like we've been using AMDGPU instead
of radeon for SI and CLK).