OpenJDK: Unset CLASSPATH before running configure.
The configure script outputs a warning and ignores CLASSPATH when it is
set. This environment variable is designed to tell the JDK where to find
user-supplied classes, which is something that isn't really necessary
for building the JDK itself. This could be used to cause issues in the
build environment (if it obeyed CLASSPATH), so let's unset it to silence
the warning.