Dolphin: add an optional external dependency on baloo-widgets.
Baloo-widgets is mentioned in CMake as being needed for the Metadata
features of Dolphin:

Make Warning at CMakeLists.txt:114 (find_package):
  By not providing "FindKF6BalooWidgets.cmake" in CMAKE_MODULE_PATH this
  project has asked CMake to find a package configuration file provided by
  "KF6BalooWidgets", but CMake did not find one.

  Could not find a package configuration file provided by "KF6BalooWidgets"
  (requested version 24.08.3) with any of the following names:

    KF6BalooWidgetsConfig.cmake
    kf6baloowidgets-config.cmake

  Add the installation prefix of "KF6BalooWidgets" to CMAKE_PREFIX_PATH or
  set "KF6BalooWidgets_DIR" to a directory containing one of the above files.
  If "KF6BalooWidgets" provides a separate development package or SDK, be
  sure it has been installed.

CMake Warning at CMakeLists.txt:131 (message):
  Baloo packages not found.  They are needed for the metadata features of
  Dolphin (including the information panel).

This might be a package that we want to add into KDE Apps later since
it's only dependency is on KF6 and it adds more core
functionality into Dolphin. KF5 used to have the needed widgets in
Baloo, but KF6 split it into kde-apps.