extra-cmake-modules: Made Qt-6 a recommended dependency.
The intructions have the -D BUILD_WITH_QT6=ON parameter. I do not if Qt-6 is required, if any Qt version is required, etc, but since the instructions say to build with Qt-6, it should be recommended at least. This causes issues in jhalfs, so this commit should fix the issue.