webkitgtk: Remove puzzling "disabled GBM" and note the patch is optional for systemd
BLFS always has GBM enabled, so "disabled GBM" is just puzzling.  We
only care the "release logs disabled" case.

In BLFS configuration, cmake leaves ENABLE_RELEASE_LOG=OFF, but the C
code has more logic:

    #if ENABLE(RELEASE_LOG)
    #define RELEASE_LOG_DISABLED 0
    #else
    #define RELEASE_LOG_DISABLED !(USE(OS_LOG) || ENABLE(JOURNALD_LOG))
    #endif

Thus on systemd we have JOURNALD_LOG=1, and release log is still
enabled.  But on sysv it's indeed disabled.  So the patch is required on
sysv, and optional for systemd.