Revert "Promote pcre2 to a required dependency for glib."
This reverts commit 744e55d03d74c300a7c117d8c285f29b7bec34eb.

If the building system automatically downloads a dependency when it's
not installed on the system, it's recommended, not required.  This rule
is well-established in the book, for example in LibreOffice most
recommended dependencies are recommended "because if they're not
installed, the build process will compile and install its own (often
older) version."

I don't see a reason to change this rule.  And even if we will change
it, we'll need a book-wide review.