grub-efi: Demote freetype to optional
With freetype at recommended, we end up a very long recommended
dependency chain: grub pulls in freetype, freetype pulls in harfbuzz,
harfbuzz pulls in graphite, graphite pulls in cmake, cmake pulls in
curl, etc.  All these are some sort of bulls*it because we actually
don't need a full freetype here, but it's not obvious to users and they
cannot skip recommended dependencies without prior knowledge.

We might add a "(harfbuzz is not needed)" for the freetype dependency
like what we've done for the pygobject dependency in
power-profiled-daemon, but to me it seems easier to just demote freetype
to optional and root out this long recommended dependency chain.  Back
to the time we added grub-efi to BLFS, if we didn't have unicode.pf2,
we'd end up having "booting up in blind mode" and the early kernel
message lines would be invisible, causing severe difficulties in boot
troubleshooting.  But that was actually a result of combination of:

1. A logic error in grub.cfg
2. The old "EFI handover" boot protocol
3. Missing unicode.pf2

Now 1 has been fixed for years, 2 has been fixed when we updated to
grub-2.12 (which prefers EFI stub over handover), and 3 alone cannot
cause the issue.

3 alone will only cause "the boot menu will be displayed using a coarse
font or in a smaller region on the screen" and to me it's completely
acceptable: nobody will really be annoyed by an ugly boot menu which
only shows up for 5 seconds on each boot.