Update to LLVM-19.1.7
printf-fortify-5.c issue is fixed, and getline_nohang.cpp is
automatically skipped.  Unfortunately we need to work around a new
regression in the test suite:
https://github.com/llvm/llvm-project/issues/123023.

The llvm-cmake tarball needs to be updated or we'll have incorrect
outputs with things like "llvm-config --version" or "clang --version".
For consistency I've updated the llvm-third-party tarball as well.
Updating them for each point version shouldn't be a large churn now as
we've decided only update LLVM at .0 and the final point release, i.e.
only twice for each stable series.

Also I've decided not to "fix up" the directory names for these two
tarballs anymore: they are extracted in the llvm source tree so the
mismatched names shouldn't be a significant issue.  Now the tarballs are
simply downloaded from upstream, renamed, then uploaded to anduin.