unzip: Clean up and update the locale issue
The patch does not work for me.  It seems taking the current locale into
the account (the result is different with LC_ALL=en_US.ISO-8859-1 and
LC_ALL=en_US.UTF-8), but the conversion is still incorrect (even if I
explicitly specify -I cp936 the result is still incorrect).

bsdunzip -I cp936 works for me anyway, so convmv is no longer needed.

Remove the wine workaround as libarchive is technically a dependency of
wine (note that in GLFS it's built from a shipped copy as a part of the
cmake package).