Seamonkey: Uncomment recommended patch listing.
The instructions before this commit would apply a patch that isn't
listed as another download. Seems it was commented out due to static
assertions, but with the latest ICU for me, everything built fine with
the patch while using system ICU. So let's uncomment the patch listing.
The patch is already in the patches repository.