mutter: Disable the test by default again and move install after test
Avoid a hard dependency on dbusmock.

We no longer need to test the package after install: the issue we worked
around doing so is already fixed by
https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3800.  So we can
use the idiomatic meson configure command to enable the test if the user
has installed dbusmock and wants to run the test, and then
libmutter-test-15.so will be installed if the test suite has been run.

Also the test of gnome-shell needs to be disabled by default again.