rpcbind: Remove patch from nowhere
After a discussion with Thomas about "unreviewed patches" I decided to
look through our patchsets.  This one seems just from nowhere.

The comment claims to fix CVE-2017-8779, but the upstream fix for this
vulnerability:

http://git.linux-nfs.org/?p=steved/libtirpc.git;a=commit;h=dd9c7cf4f8f3

does not contain the content of this patch at all.  All other distros do
not carry this patch either.

So remove it.