libpsl: Demote libidn2 and libunistring to recommended
And add icu and libidn as optional, they can be used instead of libidn2.
This package can be built with even no i18n support but I don't think
it's a good idea (due to the security implication of public suffix
list), so I kept libidn2 and libunistring at recommended instead of
optional.