Wireplumber: fix audio and video problems on SysV and systemd systems.
The problem exhibits itself during video playback as an issue where the
video will freeze on the first frame. After debugging, it turns out that
this is an issue with a conflict between Pipewire and Pulseaudio.

For systemd, we'll change to using the pipewire-pulse service.

For SysV, we'll update the pipewire-launcher script to do the following:

- Make it easier to debug by killing the previous daemons before
  starting the new ones. Pipewire refuses to run with more than one
  instance per user, and this can allow us to debug further issues
  without logging out and in to test changes.
- Wait until the pipewire service is up before we start wireplumber. On
  some slower systems, waiting for a 'sleep 1' is insufficient, and this
  leads to a race condition.
- Start the pipewire-pulse daemon.

For both init systems, we now remove the XDG autostart files for
Pulseaudio when Wireplumber is installed on the system. We also
configure /etc/pulse/client.conf to not spawn a Pulseaudio server every
time something uses the Pulseaudio libraries or utilities, as we'll run
into the same situation again.

I've also added a note into the Pulseaudio page prompting users to make
the necessary configuration adjustments if they are reinstalling or
upgrading the package on a system where Wireplumber is installed.