Update to texlive-20250308.
I've tested both the binary and source versions with:

- Ken's latex-test tarball, version 20250109
- Making the documentation for libassuan
- Building evince with the 2025 version of libkpathsea and then viewing
  a DVI file (which performs a conversion in the background)

There was also an issue with the binary which caused some applications
to not work correctly in the full install, but that was easily fixed by
referring users back to the page for setting the Texlive environment
variables (just setting PATH= is insufficient).

As Ken mentioned privately, asymptote's binary version no longer
requires libglvnd. xindy still requires the libcrypt.so.1 library, but
biber does not.

Overall this update was relatively painless!