vte: Move fast_float to recommended
Based on prior-art for libreoffice and glib, if the building system
automatically downloads a dependency when it's not installed on the
the system, it's recommended, not required.

Well I know someone doesn't like this...  So add an explicit note for
the automatical download.  I guess the explicit note should be added
for similar cases, like glib vs. pcre2.