beyond: Update DEB & RPM extraction instructions.
The tip was generally outdated and linked to scripts and code that no
longer works properly, nor compiles. .deb extraction was overengineered
and can just be extracted with `ar x` as .deb files are just ar
archives. .rpm files can be extracted via an Arch Linux package
(collection of shell scripts that utilize libarchive). If this doesn't
suffice and a GUI is preferred, File-Roller or Ark can be used instead.