libcdio: Rework i686 fix
`export CFLAGS+="-D_LARGEFILE64_SOURCE"` is really bad because when the
user does not have a CFLAGS, it sets CFLAGS=-D_LARGEFILE64_SOURCE w/o
any optimization flags in it.  Thus this package will be built
unoptimized.  And the variable is not even unset, so if the user
continues to build more autoconf-based packages in the shell, all of
them will be unoptimized.

And the problem here is the upstream doing things wrong.  The code looks
like:

    #if defined(HAVE_LSEEK64) && defined(_FILE_OFFSET_BITS) && (_FILE_OFFSET_BITS == 64)
    #define CDIO_LSEEK lseek64
    #else

But the Glibc documentations says clearly if _FILE_OFFSET_BITS == 64,
lseek will just support 64-bit offset so it does not make too much sense
to use lseek64 at all.

So we should just use lseek instead of lseek64 here.