mesa, intel-media: Mention the new Xe kernel driver
Xe is still said to be "experimental" but it seems the only choice for
Lunar Lake (Core Ultra 2xxV) and Battlemage (Arc B5xx) as at Linux-6.14.2.

Per the experience of another distro to which I contribute, enabling
both i915 and Xe and setting CONFIG_DRM_XE_FORCE_PROBE to '*' will make
the kernel use i915 if it works, and Xe if it works but i915 does not
(or the support for this GPU is also experimental in i915 anyway).  This
seems a good behavior to me: an experimental driver is better than no
driver.

If the user wants to use Xe for a GPU supported by i915 (s)he'd need to
either disable i915 or deny the GPU via CONFIG_DRM_I915_FORCE_PROBE.