libpeas: Disable Python plugin support.
On IRC earlier I received a report from a user privately who had the
non-BLFS package "gedit-plugins" installed. They reported that when they
launched Gedit, it hung with messages in the console identical to that
which you'd get when running the tests:

/usr/lib/python3.13/site-packages/gi/module.py:57: Warning: cannot register existing type 'GIRepository'
  repository = Repository.get_default()
/usr/lib/python3.13/site-packages/gi/module.py:57: Warning: g_once_init_leave_pointer: assertion 'result != 0' failed
  repository = Repository.get_default()
/usr/lib/python3.13/site-packages/gi/module.py:57: Warning: g_object_new_with_properties: assertion 'G_TYPE_IS_OBJECT (object_type)' failed
  repository = Repository.get_default()

I'd like to thank Xi first for leaving a pointer to the commit upstream
where the tests were disabled in GNOME's desktop image.

This issue is a bit tricky. It appears to be due to pygobject porting to
a new version of libgirepository without bumping it's ABI. Both
libpeas-1 and libpeas-2 are broken in the exact same way, and the issue
is tricky to solve in both of them. In the case of libpeas-1, we need
both libpeas *and* pygobject to be using the same GIRepository instance,
which isn't possible given the API removal from GLib. libpeas-2 (which
we don't carry at the moment thankfully) needs additional porting work
from gjs as well which won't be done until GNOME-49.

I was able to ascertain that from reading https://gitlab.gnome.org/GNOME/libpeas/-/issues/58

I looked at the downstream bug report mentioned at Arch, and discovered
that they disabled the Python loader for libpeas.

Let's do that here as well so we can clean out the test failure and not
run into any potential problems down the road.

Tested with both gedit and eog with all of their plugins turned on via
Preferences.