libproxy: Update the installed files and short descriptions
While I was researching how VLC's proxy configuration settings worked
(since it never links to libproxy), I discovered that it calls a utility
named 'proxy' to determine how it's supposed to access a newtork
resource.

After some more research I found out that we missed the proxy program as
well a directory, so let's add that into the installed files.

While I'm here I've also updated the short description for libproxy.so
and added a short description for the 'proxy' command