Add Adwaita fonts, etc.
1. Add the Adwaita fonts.
2. Reduce the importance of installing DejaVu.
3. Note that Cantarell was for GN~OME before 48.
4. Show the name of my font website.

In the comments I note that my host does not permit me to pay for
hosting beyond 12 months, which will be up in January.  I assume
I will no-longer be online by the time LFS-12.4 is tested.  I had
hoped the comparisons would be useful, but I start to think that
only web crawlers care.  If I have not updated the site recently
before 12.4, please either remove the link, or ask archive.org
if they can archive it.