rust: Drop stage=1
It seems the bootstrap process has been refactored, thus "stage 2"
consistently means the code produced by the stage-1 compiler now.  On
the contrary, installing stage-1 components is now incorrect as they are
built with stage 0 (i.e. a previous release of rust) and so not settled
down.

Notably, this issue breaks "cargo clippy" as stage-1 clippy needs
stage-1 librustc_driver.so that we do not install (only the stage-2
librustc_driver.so is installed as a part of stage-2 compiler, and the
two librustc_driver.so's have different SONAME as the ABI of rust isn't
stablized yet).  Thus "ninja test" of snapshot will fail.

Remove the stage overrides to align with the upstream change and also
simplify the instructions.