Fix building Texlive with GCC 15.
This isn't the way I really wanted to do it, but the source tree is far
too complex to fix manually. Upstream has made a lot of process, but
core components such as lualatex and cjisutils and other related
components are not able to be compiled with C23 at this time.

I'll start checking weekly for updates at https://github.com/TeX-Live/texlive-source

I picked up 14 patches from here and tried applying them to my tree but
it broke at lualatex as it hasn't been updated yet.