Protobuf: Don't delete libupb.a.
It's been quite a bit since the discussion on the mailing list. This
needed to get done sooner rather than later. This is a temporary
solution but we need at least *a* solution to fix the build with Qt-6.
Implement solution a now, bikeshed later and come up with a better one.
Sorry for making a change now but it needs to get done so users aren't
left wondering what to do and wonder why the build is broken.

Also apologies for the tone is this git commit message.