OpenJDK: Remove *.debuginfo and restore MAKEFLAGS.
*.debuginfo takes over half of the installed size of JDK SDKs, so
providing an command that removes them helps a lot. It also documents a
technique we do for providing distributions of OpenJDK. MAKEFLAGS is
also unset, and will remain unset until the user sets it again. This can
lead to a reader forgetting it and having much slower build times in
proceeding packages. Thus, we now store MAKEFLAGS in a hold variable so
that the original value can be restored after compilation, and that the
hold variable is properly unset afterwords.