Thunderbird: Account for newer 140.0esr release.
Hopefully this should be the last shadow update for the 140.0esr update.
All this commit does is change the tarball size and md5sum. Between
shadow updates, nothing meaningful changes, just timestamp information
and commit hashes. Functional code was not changed. Thus, there is no
need to do a rebuild. Strikes me as odd that the size increased by 2M
from the original release, but I suppose it doesn't matter much as long
as the diff shows that nothing meaningful changed.

Either way, we should be good now.