GDM: Fix starting on SysV when only X11 is usable.
This problem has been around for quite a while. Users who have Xorg as
SUID-root aren't impacted, but users who have the default configuration
are.

The default configuration for GDM is to use Wayland. This is well
supported on SysV and works great, unless you are running LFS in a
hypervisor such as VMWare, or if you are using an NVIDIA GPU with the
proprietary driver.

There were two issues that were preventing this from working correctly.
One was the PAM configuration being slightly incorrect, which caused
elogind to be optionally loaded... preventing a takeover of the virtual
terminal. elogind is also only running on SysV once something activates
it, and we rely off PAM to do this for us. Other distributions don't
need to pass the -D initial-vt option because their copy of elogind is
started via a bootscript. To work around this for now and prevent making
major changes right before package freeze, let's just set the initial VT
to 7.

I've tested this on a SysV VM in VMWare Workstation 17, as well as with
a system with an NVIDIA RTX 3060 in it.