webkitgtk: Add more info about the RAM usage issue
The 16 GiB threshold is just too vauge.  I have a laptop with 18 logical
cores and 32 GiB RAM, and obviously it'll OOM if I don't limit the
number of build jobs.  But it's not so obvious to normal readers IMO.

And the 32-bit issue is technically different from the generic (64-bit)
issue.  On 64-bit you can always add swap to overcome RAM shortage
(though the build will be stupidly slow), but on 32-bit it'll just fail
even if you have 1 TB of swap (note that HIGHMEM64G is removed from the
kernel so you'll no longer be able to use more than 4 GiB of real RAM
with a 32-bit kernel anyway) and degrading from -O3 to to -O2 is the
only option.

Reword the <important> box to make those issues more clear.