PostgreSQL: minor tweaks to the instructions, and tag it.
The instructions by default run 'make install-docs'. The PostgreSQL
tarball does not appear to come with the documentation pregenerated
anymore, and as a result the build fails due to the missing SGML stack
and fop, which are currently optional.

To work around this, let's break 'make install-docs' out, and then
mention to users that they can run it if they have the optional
dependencies installed for generating the documentation.