llvm: Document new test failures
atomic.cpp needs to be built with -D_GLIBCXX_NO_ASSERTIONS with
libstdc++ from GCC-15 to avoid referencing __glibcxx_assert_fail, see
https://github.com/llvm/llvm-project/commit/6beb6f34bc8e.

Quote from
https://github.com/google/sanitizers/wiki/AddressSanitizerCallStack:

   malloc and free happen relatively frequently and it is important to
   unwind the call stack fast. AddressSanitizer uses a simple unwinder
   that relies on frame pointers.

But in LFS Glibc is built with -O2 which implies -fomit-frame-pointer.
Thus this unwinder just cannot work reliably on it (i.e. the trace past
main() may be incorrect).  This makes an extra error line from
llvm-symbolizer into the stderr in onprint.cpp, causing it different
from the output file created by the sanitizer callback and breaking
this test.

I guess the ultimate solution to the second issue is SFrame, but
Glibc-2.42 isn't built with SFrame enabled either (because some
severe SFrame bugs in Binutils was detected at the last moment of
Glibc-2.42 development cycle) and the sanitizer unwinder still needs to
be adapted to use it.