Emergency Security Updates for BLFS 12.4.
This includes updates for the following packages:

Qt6/QtWebEngine 6.9.1 to 6.9.2
krb5 1.22.0 to 1.22.1
udisks2 2.10.1 to 2.20.2

This also updates Libreoffice to 25.8.1.1 to fix a critical crasher bug
and also problems with bolding/unbolding text.

Yesterday I retested all dependents of these packages. Testing included
basic functionality tests for all packages, as well as lots of testing
for Falkon, Plasma, LXQt, and Libreoffice. In terms of retesting I have
also rebuilt *all* packages affected, 198 yesterday in total across two
systems.

There were several text changes that I made to Kerberos. These are all
grammar/clarification fixes.

The patches for Qt6 and the sed for udisks are no longer needed.

There will be several commits after this one containing a fix for Guile,
and for several text tweaks and fixes related to Kerberos support.