appstream: Freeze 12.4 release into the LFS metainfo file
I don't know how would we model the development branch.  IMO we cannot
use &version; and &pubdate; here because they are corresponding to BLFS,
not LFS.

One way may be creating an org.linuxfromscratch.lfs.releases.xml as a
part of the LFS book and refer to it using an URL in this file.  But the
specification says the external file must be also installed onto the
system, see
https://www.freedesktop.org/software/appstream/docs/sect-Metadata-Releases.html#spec-releases-location.

So to me the easiest way is simply not to show the development branch
here.