rust: Make the recommended dependency handling consistent
We always assume the recommended dependencies are satisified in the book
instruction (note that we are already assuming LLVM is installed in the
very same page), so we don't need to have the [ ... ] || stuff for
*_USE_PKG_CONFIG=1, but we should document them in command explanations.

Reword a note which seemed to mislead users into believing the book
instructions don't need a change for omitting optional dependency(ies).

By the way, update the info about Glibc stripping issue.