gtk4: Mention the UDMABUF kernel configuration
In the past I refused to mention it because it was only needed by the
test suite, and it was not accessible by non-root users.  But now it's
also used by the GTK4 library itself (though I don't know which specific
application needs it), and the default udev rules of systemd-256 allow
the non-root users to access it.

So mention it now and enable the tests using it.