dav1d; SVT-AV1: BLFS standards.
I ported the dav1d and SVT-AV1 pages from GLFS, and standards differ
between that and this book. This commit ensures the pages better fall in
line with what is done in BLFS.

Another change for telling the user they need an Internet connection to
run the test suite for SVT-AV1.