FFmpeg: Dependency order, missing (, and comment.
The comment in particular is in regards to VPDAU and how we link to it
for NVIDIA systems. Mesa-25.3 will remove VDPAU drivers and we don't
provide NVIDIA proprietary. Either we demote to Optional or remove the
section. It will be a bit before we have to make a decision, however.

I will be thinking about this in the meantime.