librsvg: Demote gdk-pixbuf to recommended
It's possible to build librsvg w/o gdk-pixbuf using -D pixbuf=disabled
-D introspection=disabled, but doing so would break ABI and lose the
introspection data.

This was an attempt to break the circular dependency gdk-pixbuf ->
glycin -> librsvg -> gdk-pixbuf, but then it turns out to be easier to
break the circle at gdk-pixbuf instead.  That will be done in the next
commit.