gdk-pixbuf, librsvg: Break the circular dependency
As gdk-pixbuf has the minimal SBU value among the packages in the
circle, and it's trivial to adapt the build instruction to make the
build configuration sane with and without glycin, break the circle here.
Tell the users (and jhalfs) to build gdk-pixbuf first w/o glycin: the
first build will be unable to load most image formats but it should be
enough to build librsvg.

There might be still some loose ends (perhaps librsvg test suite can be
unhappy) and I'll tie them (if any) before 12.5 release.