Seamonkey: Adjust ICU line break mapping fixing procedure.
Currently, the for loop would break on ICU-77 systems as it's meant for
78.1. This essentially reverts the change in the Seamonkey update I
pushed to this repository where the 47 number got increased to 48 to
adapt for ICU-78. Instead, there is an additional sed and more
documentation, adding on that there is another line break mapping fix
for ICU-78 in particular. We want to make sure older ICU versions
installed can still get the newest Seamonkey while ensuring the newest
ICU can still be used. This setup is not ideal but is probably the best
solution without breaking the procedure for previous ICU versions.