Revert "Seamonkey: Adjust ICU line break mapping fixing procedure."
This reverts commit 53348e907b14fdb02177d8d953019cb48f694ea7.

xry111 sent an email covering this commit, explaining how we do not
support previous versions of ICU. I believe this is a different
situation and we should think about how to approach ICU in the future.
Nonetheless, the SA has been edited to note how the sed change is
incompatible, so this commit makes it so it is indeed incompatible and
shortens the instructions a bit.

If any editors feel there is more work to be done in regards to this
change, send an email for this commit or talk about it in IRC so things
can be done right.