gdk-pixbuf: Pass -D glycin=disabled for no-glycin build
Because there exists a circular recommended dependency (gdk-pixbuf ->
glycin -> librsvg -> gdk-pixbuf), we need to build one of the packages
twice.  The gdk-pixbuf package has the smallest SBU so this is where we
break the loop.

I initially thought the default of glycin is "auto" so the first build
should automatically work, but it turns out gdk-pixbuf meson.build has
some magic to set it "enabled" on Linux.  Use some shell magic as our
own "auto" logic here.