libsoup3: Fix several CVEs and a potential deadlock when using gstreamer.
The CVEs fixed include:

- CVE-2025-12105
- CVE-2025-4948
- CVE-2025-4945
- CVE-2025-4969
- CVE-2025-4476
- CVE-2025-32914
- CVE-2025-32908
- CVE-2025-32907

As well as one which has not been assigned a CVE
(https://gitlab.gnome.org/GNOME/libsoup/-/issues/446)

There are still more unresolved, but they have either no upstream fix
(in the case of CVE-2025-9901), require a major API change which isn't
reliable with our version of libsoup (and where a patch is unavailable
for the 3.6 branch - CVE-2025-32049), or are better fixed through
another package (CVE-2025-4035). In addition, there are four more which
will not be made public until their disclosure dates in December and
January.

Let's get us up to date for now though with what we have access to...