Update to KDE Gear 25.08.3
This commit inclues all the packages in BLFS Chapter 31
(KDE Frameworks 6 Based Applications) as well as
kate
falkon
kwave

Two packages in Chapter 31 are based on Plasma packages not yet in the book
(plasma-activities and plasma-activities-stats) and the plasma ENTITY has not
yet been updated from 6.4.4 to 6.5.2, but it will be updated very soon.  The
internl stats for those packages have been updated.