SDL: Multiple changes.
- Added SDL-3.2.26.
- Switched from SDL2-2.32.10 to sdl2-compat-2.32.58.
- Removed sdl12-compat-1.2.70.

Testing still needs to be done and packages that can use SDL3 but the
book says SDL2 need to have their pages edited. This is where the
benefit of this branch comes in. It will undergo testing and minor
changes until it's ready. It shall not be merged until everything is
confirmed to work.

I will privately mark off packages that have been confirmed and will
occasionally update as needed at
https://wiki.linuxfromscratch.org/blfs/ticket/20958.