Sphinx: Document test failures after some research.
These appear to be harmless at a glance, but we'll have to keep an eye
on packages that use Sphinx.

Five of the failures are related to docutils-0.22.x. The other one is
related to Python 3.14, but the fix for it requires a large refactor of
the autodoc code - so it is best to wait on a new version of Sphinx to
fix those for now (unless they start causing problems... if one of us
encounters issues with it, I will try to create a backport or a git
snapshot).