commonmark: Remove an unnecessary dependency on flake8
The setup.py file notes that it's required for tests. However, I've
tested the test suite both with and without it, and there are no changes
in the tests run when it's installed.

Since we already have text describing just the hypothesis module, let's
pull flake8 from the dependencies list.