ply: Add the test suite back to the book, and fix it with Python 3.12+
This allows some basic tests of the module to be run. One suite is still
broken, testcpp.py. When running that suite, multiprocessing errors are
emitted. However, this still gives us peace of mind that the features
tested by contributors to ply are still functional, even if it hasn't
been touched for over 2 years by upstream.

This is a combination of Arch's patch for Python 3.11 support, and
upstream PR#307.